requirements.txt
aiogram>=3.17.0
redis>=5.2.1
sqlalchemy>=2.0.37
asyncpg>=0.30.0
pydantic-settings>=2.7.1
python-docx>=1.1.2
alembic>=1.14.0
structlog>=25.1.0
python-dotenv>=1.0.1
cryptography>=44.0.0
